requirements/prod.txt
aio-pika==9.6.1
aiohttp==3.9.3
alembic==1.13.1
asyncpg==0.29.0
fastapi==0.115.12
numpy>=1.26,<3
pydantic==2.10.6
python-dotenv==1.2.2
httpx==0.28.1
python-multipart==0.0.20
SQLAlchemy==2.0.27
uvicorn==0.34.2
